requirements.txt
torch>=2.3.0
torchvision>=0.18.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
jsonschema>=4.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
transformers>=4.30.0
diffusers>=0.18.0
accelerate>=0.20.0

